Refactor triggers and add abac
