+++ studio/components/interfaces/Database/Triggers/TriggersList/TriggersList.tsx
@@ -1,0 +1,122 @@
+import { FC } from 'react'
+import { includes, uniqBy, map as lodashMap } from 'lodash'
+import { observer } from 'mobx-react-lite'
+import * as Tooltip from '@radix-ui/react-tooltip'
+import { Typography, Button, Input, IconSearch, IconLoader } from '@supabase/ui'
+import { PermissionAction } from '@supabase/shared-types/out/constants'
+
+import { checkPermissions, useStore } from 'hooks'
+import SchemaTable from './SchemaTable'
+import AlphaPreview from 'components/to-be-cleaned/AlphaPreview'
+import ProductEmptyState from 'components/to-be-cleaned/ProductEmptyState'
+
+const TriggersList: FC<any> = ({
+  filterString,
+  setFilterString = () => {},
+  createTrigger = () => {},
+  editTrigger = () => {},
+  deleteTrigger = () => {},
+}) => {
+  const { meta } = useStore()
+  const triggers = meta.triggers.list()
+  const filteredTriggers = triggers.filter((x: any) =>
+    includes(x.name.toLowerCase(), filterString.toLowerCase())
+  )
+  const filteredTriggerSchemas = lodashMap(uniqBy(filteredTriggers, 'schema'), 'schema')
+  const canCreateTriggers = checkPermissions(PermissionAction.TENANT_SQL_ADMIN_WRITE, 'triggers')
+
+  if (meta.triggers.isLoading) {
+    return (
+      <div className="flex h-full w-full items-center justify-center space-x-2">
+        <IconLoader className="animate-spin" size={14} />
+        <Typography.Text>Loading triggers...</Typography.Text>
+      </div>
+    )
+  }
+
+  if (meta.triggers.hasError) {
+    return (
+      <Typography.Text className="px-6 py-4" type="danger">
+        <p>Error connecting to API</p>
+        <p>{`${meta.triggers.error?.message ?? 'Unknown error'}`}</p>
+      </Typography.Text>
+    )
+  }
+
+  return (
+    <>
+      {triggers.length == 0 ? (
+        <div className="flex h-full w-full items-center justify-center">
+          <ProductEmptyState
+            title="Triggers"
+            ctaButtonLabel="Create a new trigger"
+            onClickCta={() => createTrigger()}
+          >
+            <AlphaPreview />
+            <p className="text-scale-1100 text-sm">
+              A PostgreSQL trigger is a function invoked automatically whenever an event associated
+              with a table occurs.
+            </p>
+            <p className="text-scale-1100 text-sm">
+              An event could be any of the following: INSERT, UPDATE, DELETE. A trigger is a special
+              user-defined function associated with a table.
+            </p>
+          </ProductEmptyState>
+        </div>
+      ) : (
+        <div className="w-full space-y-4 py-4">
+          <div className="flex items-center justify-between px-6">
+            <Input
+              placeholder="Filter by name"
+              size="small"
+              icon={<IconSearch size="tiny" />}
+              value={filterString}
+              onChange={(e) => setFilterString(e.target.value)}
+            />
+            <Tooltip.Root delayDuration={0}>
+              <Tooltip.Trigger>
+                <Button disabled={!canCreateTriggers} onClick={() => createTrigger()}>
+                  Create a new trigger
+                </Button>
+              </Tooltip.Trigger>
+              {!canCreateTriggers && (
+                <Tooltip.Content side="bottom">
+                  <Tooltip.Arrow className="radix-tooltip-arrow" />
+                  <div
+                    className={[
+                      'bg-scale-100 rounded py-1 px-2 leading-none shadow',
+                      'border-scale-200 border',
+                    ].join(' ')}
+                  >
+                    <span className="text-scale-1200 text-xs">
+                      You need additional permissions to create triggers
+                    </span>
+                  </div>
+                </Tooltip.Content>
+              )}
+            </Tooltip.Root>
+          </div>
+          {filteredTriggers.length <= 0 && (
+            <div className="dark:border-dark mx-auto flex max-w-lg items-center justify-center space-x-3 rounded border p-6 shadow-md">
+              <Typography.Text>No results match your filter query</Typography.Text>
+              <Button type="outline" onClick={() => setFilterString('')}>
+                Reset filter
+              </Button>
+            </div>
+          )}
+          {filteredTriggerSchemas.map((schema: any) => (
+            <SchemaTable
+              key={schema}
+              filterString={filterString}
+              schema={schema}
+              editTrigger={editTrigger}
+              deleteTrigger={deleteTrigger}
+            />
+          ))}
+        </div>
+      )}
+    </>
+  )
+}
+
+export default observer(TriggersList)
